Refactor AWSClient to extract Json, XML clients, and the presign method
